requirements.txt
annotated-types==0.6.0; python_version >= '3.8'
anyio==3.7.1; python_version >= '3.7'
click==8.1.7; python_version >= '3.7'
fastapi==0.104.1
h11==0.14.0; python_version >= '3.7'
idna==3.4; python_version >= '3.5'
pydantic==2.4.2
pydantic-core==2.10.1; python_version >= '3.7'
python-dotenv==1.0.0
sniffio==1.3.0; python_version >= '3.7'
starlette==0.27.0; python_version >= '3.7'
typing-extensions==4.8.0; python_version >= '3.8'
uvicorn==0.24.0